TASK: Move object validation out of PersistenceManager
This moves the code used to validate objects using the onFlush event
of Doctrine out of the PersistenceManager and into it's own class.

The listener is then added using Flow settings, instead of PHP code
inside PersistenceManager.initializeObject().
