+++ Neos.Flow/Classes/Persistence/Doctrine/PersistenceManager.php
@@ -19,13 +19,10 @@ use Neos\Flow\Annotations as Flow;
 use Neos\Flow\Log\SystemLoggerInterface;
 use Neos\Flow\Persistence\AbstractPersistenceManager;
 use Neos\Flow\Persistence\Exception\KnownObjectException;
-use Neos\Flow\Persistence\Exception\ObjectValidationFailedException;
 use Neos\Flow\Persistence\Exception as PersistenceException;
 use Neos\Flow\Persistence\Exception\UnknownObjectException;
-use Neos\Flow\Reflection\ClassSchema;
 use Neos\Utility\ObjectAccess;
 use Neos\Flow\Reflection\ReflectionService;
-use Neos\Utility\TypeHandling;
 use Neos\Flow\Validation\ValidatorResolver;
 
 /**
@@ -59,84 +56,6 @@ class PersistenceManager extends AbstractPersistenceManager
      * @var ReflectionService
      */
     protected $reflectionService;
-
-    /**
-     * Initializes the persistence manager, called by Flow.
-     *
-     * @return void
-     */
-    public function initializeObject()
-    {
-        $this->entityManager->getEventManager()->addEventListener([\Doctrine\ORM\Events::onFlush], $this);
-    }
-
-    /**
-     * An onFlush event listener used to validate entities upon persistence.
-     *
-     * @param \Doctrine\ORM\Event\OnFlushEventArgs $eventArgs
-     * @return void
-     */
-    public function onFlush(\Doctrine\ORM\Event\OnFlushEventArgs $eventArgs)
-    {
-        $unitOfWork = $this->entityManager->getUnitOfWork();
-        $entityInsertions = $unitOfWork->getScheduledEntityInsertions();
-
-        $validatedInstancesContainer = new \SplObjectStorage();
-        $knownValueObjects = [];
-        foreach ($entityInsertions as $entity) {
-            $className = TypeHandling::getTypeForValue($entity);
-            if ($this->reflectionService->getClassSchema($className)->getModelType() === ClassSchema::MODELTYPE_VALUEOBJECT) {
-                $identifier = $this->getIdentifierByObject($entity);
-
-                if (isset($knownValueObjects[$className][$identifier]) || $unitOfWork->getEntityPersister($className)->exists($entity)) {
-                    unset($entityInsertions[spl_object_hash($entity)]);
-                    continue;
-                }
-
-                $knownValueObjects[$className][$identifier] = true;
-            }
-            $this->validateObject($entity, $validatedInstancesContainer);
-        }
-
-        ObjectAccess::setProperty($unitOfWork, 'entityInsertions', $entityInsertions, true);
-
-        foreach ($unitOfWork->getScheduledEntityUpdates() as $entity) {
-            $this->validateObject($entity, $validatedInstancesContainer);
-        }
-    }
-
-    /**
-     * Validates the given object and throws an exception if validation fails.
-     *
-     * @param object $object
-     * @param \SplObjectStorage $validatedInstancesContainer
-     * @return void
-     * @throws ObjectValidationFailedException
-     */
-    protected function validateObject($object, \SplObjectStorage $validatedInstancesContainer)
-    {
-        $className = $this->entityManager->getClassMetadata(get_class($object))->getName();
-        $validator = $this->validatorResolver->getBaseValidatorConjunction($className, ['Persistence', 'Default']);
-        if ($validator === null) {
-            return;
-        }
-
-        $validator->setValidatedInstancesContainer($validatedInstancesContainer);
-        $validationResult = $validator->validate($object);
-        if ($validationResult->hasErrors()) {
-            $errorMessages = '';
-            $errorCount = 0;
-            $allErrors = $validationResult->getFlattenedErrors();
-            foreach ($allErrors as $path => $errors) {
-                $errorMessages .= $path . ':' . PHP_EOL;
-                foreach ($errors as $error) {
-                    $errorCount++;
-                    $errorMessages .= (string)$error . PHP_EOL;
-                }
-            }
-            throw new ObjectValidationFailedException('An instance of "' . get_class($object) . '" failed to pass validation with ' . $errorCount . ' error(s): ' . PHP_EOL . $errorMessages, 1322585164);
-        }
-    }
 
     /**
      * Commits new objects and changes to objects in the current persistence
