+++ Neos.Flow/Configuration/Settings.yaml
@@ -353,7 +353,10 @@ Neos:
         # Doctrine ORM Lifecycle events and their subscribers and listeners.
         # See Flow manual, chapter "Persistence" for further information on how to subscribe to Doctrine events.
         eventSubscribers: []
-        eventListeners: []
+        eventListeners:
+          -
+            events: ['onFlush']
+            listener: 'Neos\Flow\Persistence\Doctrine\ObjectValidationListener'
 
         secondLevelCache: []
